% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{BinAnnotationPlot}
\alias{BinAnnotationPlot}
\title{Heatmap of bin-annotation overlap}
\usage{
BinAnnotationPlot(seu, cell.type.calls = "celltype.l2")
}
\arguments{
\item{seu}{A seurat object with binning IDs in the "bins" column of meta.data}

\item{cell.type.calls}{Name of meta.data column containing cell type calls on which to visualize overlap}
}
\value{
A heatmap depicting cell type calls within each bin
}
\description{
Heatmap of bin-annotation overlap
}
\examples{
\dontrun{
BinAnnotationPlot(seu)
}
}
