% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{BinDatasets}
\alias{BinDatasets}
\title{Align Datasets through binning}
\usage{
BinDatasets(
  seu,
  split.by = "time.orig",
  dims = 1:50,
  coarse_cell_types = NULL,
  sigtest_cell_types = NULL,
  snn.reduction = "pca",
  anchor_score_threshold = 0.5,
  optim_quan.threshold = 0.1,
  optim_k.unique = NULL,
  verbose = F
)
}
\arguments{
\item{split.by}{Name of meta.data column containing the factor by which the dataset will be split}

\item{dims}{Dimensions of reduction to use as input for neighbor graph calculation}

\item{coarse_cell_types}{Name of meta.data column containing cell type calls for binning. Each set of cell type calls will be binned separately.}

\item{sigtest_cell_types}{Name of meta.data column containing the cell type labels to use for connectivity testing}

\item{snn.reduction}{Name of reduction used as input to building the SNN}

\item{anchor_score_threshold}{Anchor pairs scoring below this threshold will be discarded.}

\item{optim_quan.threshold}{Percentage of poor connectivity bins to remove on each iteration of connectivity optimization}

\item{optim_k.unique}{Mean number of datasets represented in each bin at which connectivity optimization will be considered complete}

\item{seuObj}{A seurat object}
}
\value{
A binned Seurat object with bin assignments in the "bins" column of the meta.data slot
}
\description{
Align Datasets through binning
}
\examples{
seu <- BinDatasets(seu, split.by = "time")
}
