% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_diff.R
\name{BinPostHoc}
\alias{BinPostHoc}
\title{Title}
\usage{
BinPostHoc(
  seu,
  bin_pvals,
  interaction_graphs,
  split.by = "time.orig",
  cell.type.calls = "celltype.l2"
)
}
\arguments{
\item{cell.type.calls}{}
}
\value{

}
\description{
Title
}
