% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{BuildCompHeatmap}
\alias{BuildCompHeatmap}
\title{Title}
\usage{
BuildCompHeatmap(
  interaction_graphs,
  seu,
  name = "var_results",
  ident.1 = NULL,
  ident.2 = "pre",
  row.k = NULL,
  col.k = NULL
)
}
\arguments{
\item{col.k}{}
}
\value{

}
\description{
Title
}
