% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{BuildGranHeatmap}
\alias{BuildGranHeatmap}
\title{Title}
\usage{
BuildGranHeatmap(
  interaction_graphs,
  seu,
  name = "var_results",
  row.k = NULL,
  col.k = NULL,
  cell.type.calls = "celltype.l2"
)
}
\arguments{
\item{cell.type.calls}{}
}
\value{

}
\description{
Title
}
