% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_graph.R
\name{BuildPriorInteraction}
\alias{BuildPriorInteraction}
\title{Title}
\usage{
BuildPriorInteraction(
  object,
  assay = "SCT",
  slot = "data",
  database = "OmniPath",
  ligands = NULL,
  recepts = NULL,
  specific = F,
  ranked_genes = NULL,
  correct.depth = T,
  graph_name = "prior_interaction"
)
}
\arguments{
\item{graph_name}{}
}
\value{

}
\description{
Title
}
