% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{BuildSingleHeatmap}
\alias{BuildSingleHeatmap}
\title{Title}
\usage{
BuildSingleHeatmap(
  interaction_graph,
  seu,
  name = "var_results",
  row.k = NULL,
  col.k = NULL,
  cell.type.calls = "celltype.l2",
  cell.cols = NULL
)
}
\arguments{
\item{cell.cols}{}
}
\value{

}
\description{
Title
}
