% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_graph.R
\name{BuildWeightedInteraction}
\alias{BuildWeightedInteraction}
\title{Title}
\usage{
BuildWeightedInteraction(
  object,
  nichenet_results = late1.nnr,
  assay = "SCT",
  slot = "data",
  pearson.cutoff = 0.1,
  scale.factors = c(1.5, 3),
  database = "OmniPath",
  ligands = NULL,
  recepts = NULL,
  correct.depth = T,
  graph_name = "weighted_interaction"
)
}
\arguments{
\item{graph_name}{}
}
\value{

}
\description{
Title
}
