% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactome.R
\name{BuildWeightedInteraction_test}
\alias{BuildWeightedInteraction_test}
\title{Title}
\usage{
BuildWeightedInteraction_test(
  object,
  nichenet_results = late1.nnr,
  assay = "SCT",
  slot = "data",
  pearson.cutoff = 0.1,
  scale.factors = c(1.5, 3),
  database = "fantom5",
  ligands = NULL,
  recepts = NULL
)
}
\arguments{
\item{recepts}{}
}
\value{

}
\description{
Title
}
