% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{CCIMDimPlot}
\alias{CCIMDimPlot}
\title{Title}
\usage{
CCIMDimPlot(
  ccim,
  seu,
  features = NULL,
  type_plot = c("sender", "receiver"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
