% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{check_completion}
\alias{check_completion}
\title{Helper function to check completion of bins}
\usage{
check_completion(nbs)
}
\arguments{
\item{nbs}{}
}
\value{
logical indicating if binning is complete
}
\description{
Helper function to check completion of bins
}
