% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_testing.R
\name{crGeneSig}
\alias{crGeneSig}
\title{Create single-cell gene signature}
\usage{
crGeneSig(seu, variant_features = NULL, quantile_threshold = 0.1)
}
\arguments{
\item{seu}{A seurat object}

\item{variant_features}{Character vector of variable genes returned by IDVariantGenes}

\item{quantile_threshold}{Genes a distance less than this threshold will be considered part of the gene signature}
}
\value{
A list n cells long with nearest gene-cell distances for each cell
}
\description{
Create single-cell gene signature
}
\examples{
\dontrun{
ranked_genes <- crGeneSig(seu, variant_features = var_genes)
}
}
