% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactome.R
\name{ExploreClusters}
\alias{ExploreClusters}
\title{Title}
\usage{
ExploreClusters(
  seu,
  ranked_genes,
  send_cells = NULL,
  rec_cells = NULL,
  clusters = NULL,
  send_cluster = NULL,
  rec_cluster = NULL,
  name = NULL,
  cell.type.calls = "celltype.l2",
  ident.label = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
