% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactome.R
\name{ExploreInteractome}
\alias{ExploreInteractome}
\title{Title}
\usage{
ExploreInteractome(seu, ranked_genes, send_cells, rec_cells, name = NULL)
}
\arguments{
\item{name}{}
}
\value{

}
\description{
Title
}
