% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_programs.R
\name{FindAllModules}
\alias{FindAllModules}
\title{Title}
\usage{
FindAllModules(
  seu,
  group.by = NULL,
  n.replicate = 10000,
  min.members = 1,
  return.mats = F,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
