% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuits.R
\name{FormatCircuits}
\alias{FormatCircuits}
\title{Stitch multi-timepoint circuits together}
\usage{
FormatCircuits(all_circuits = NULL, all.tps = NULL)
}
\arguments{
\item{all.tps}{}
}
\value{

}
\description{
Stitch multi-timepoint circuits together
}
