% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_graph.R
\name{GenerateCCIM}
\alias{GenerateCCIM}
\title{Title}
\usage{
GenerateCCIM(
  object,
  assay = "SCT",
  slot = "data",
  database = "OmniPath",
  ligands = NULL,
  recepts = NULL,
  senders = NULL,
  receivers = NULL
)
}
\arguments{
\item{receivers}{}
}
\value{

}
\description{
Title
}
