% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_edges.R
\name{GenerateCellSignature}
\alias{GenerateCellSignature}
\title{Generate single-cell gene signature for ligand activity ranking}
\usage{
GenerateCellSignature(seu, variant_genes, dq = 0.05)
}
\arguments{
\item{seu}{A seurat object}

\item{variant_genes}{character vector of genes to use for gene-signature calculation. This list of genes should comprise genes that are variable across the biological axis of interest.
Eg. for single datasets, this may simply be the dataset's HVGs. For a longitudinal dataset, these may be genes that change over time. Use \code{IDVariantGenes} to identify genes that vary across time, between groups, etc.}

\item{dq}{numeric. Distance quantile. Genes that are further away than this quantile threshold will not be considered part of a cell's gene signature and not used for ligand activity prediction}
}
\value{
Returns a matrix where columns are cells, rows are genes, and values are binary corresponding to if that gene appears in that cell's gene signature
}
\description{
Generate single-cell gene signature for ligand activity ranking
}
\references{
Cortal, et al. Nat Biotech (2021)
}
