% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactome.R
\name{GenerateInteractome}
\alias{GenerateInteractome}
\title{Title}
\usage{
GenerateInteractome(
  seu,
  ranked_genes = ranked_genes,
  cluster_results = clusters,
  send_cluster = NULL,
  rec_cluster = NULL,
  ident.label = NULL,
  use_clusters = F,
  send_cells = NULL,
  rec_cells = NULL,
  cell.type.calls = "celltype.l2",
  assay.use = "SCT",
  slot.use = "data",
  group.by = "time.orig",
  database = "OmniPath"
)
}
\arguments{
\item{database}{}
}
\value{

}
\description{
Title
}
