% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_programs.R
\name{InteractionModules}
\alias{InteractionModules}
\title{Title}
\usage{
InteractionModules(
  object,
  assay = "SCT",
  slot = "data",
  database = "OmniPath",
  ligands = NULL,
  recepts = NULL,
  iterate.threshold = 500,
  n.iterate = NULL,
  specific = F,
  ranked_genes = NULL,
  return.mat = F,
  softPower = 1,
  min.size = 5,
  plot.mods = F,
  tree.cut.quantile = 0.4
)
}
\arguments{
\item{tree.cut.quantile}{}
}
\value{

}
\description{
Title
}
