% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_programs.R
\name{InteractionProgramSignificance}
\alias{InteractionProgramSignificance}
\title{Title}
\usage{
InteractionProgramSignificance(ip_data, n.replicate = 10000, min.members = 1)
}
\arguments{
\item{min.members}{}
}
\value{

}
\description{
Title
}
