% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_programs.R
\name{IPCellTypeSummary}
\alias{IPCellTypeSummary}
\title{Identify most highly expressed interaction programs by cell type}
\usage{
IPCellTypeSummary(seu, group.by)
}
\arguments{
\item{seu}{A seurat object where interaction program expression has been scored by \code{ScoreInteractionPrograms}}

\item{group.by}{\code{meta.data} column corresponding to the cell type or cluster annotations to use for aggregating Interaction program expression values}
}
\value{
A data.frame with average interaction program expression score for each cell type combination.
}
\description{
Identify most highly expressed interaction programs by cell type
}
