% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{IPFeaturePlot}
\alias{IPFeaturePlot}
\title{Plot expression of interaction programs}
\usage{
IPFeaturePlot(seu, ip, cols = c("grey90", "blue", "orangered3"), order = T)
}
\arguments{
\item{seu}{A seurat object that has interaction program expression information stored in the meta.data slot as the output of \code{ScoreInteractionPrograms}}

\item{ip}{The name of the interaction program to plot}

\item{cols}{Character vector of colors to use for plotting}

\item{order}{Logical. Order points so highest expressers are on top?}
}
\value{

}
\description{
Plot expression of interaction programs
}
