% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LoadRawLR}
\alias{LoadRawLR}
\title{Load unformatted ligand-receptor database}
\usage{
LoadRawLR(species = "human", database = "OmniPath")
}
\arguments{
\item{species}{character. Name of species from which to load ligand-receptor databases. One of: "human", "mouse", "rat". Default: "human"}

\item{database}{Name of ligand-receptor database to use. Default: "OmniPath"
When species is "human", one of: OmniPath, CellChatDB, CellPhoneDB, Ramilowski2015, Baccin2019, LRdb, Kirouac2010, ICELLNET, iTALK, EMBRACE, HPMR, Guide2Pharma, connectomeDB2020, talklr, CellTalkDB
When species is "mouse" or "rat", only "OmniPath" is supported.
To pass a custom ligand-receptor database to this function, set database = "custom"}
}
\value{
Returns a dataframe containing full information for the specified LR database. If species==NULL, will return all LR databases in the package. If database==NULL, will return all databases for the given species.
}
\description{
Load unformatted ligand-receptor database
}
\references{
Turei, et al. Molecular Systems Biology (2021); Raredon, et al. bioRxiv (2021)
}
