% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_graph.R
\name{MapMetaData}
\alias{MapMetaData}
\title{Map metadata from Seurat object to CCIM object}
\usage{
MapMetaData(ccim_seu, seu, columns_map = NULL)
}
\arguments{
\item{ccim_seu}{Seurat object generated by GenerateCCIM that contains a cell-cell interaction matrix}

\item{seu}{Parent Seurat object from which the cell-cell interaction matrix was generated}

\item{columns_map}{Character vector of names of metadata columns to map from parent Seurat object}
}
\value{
Returns a Seurat object with metadata from each sender and receiver cell mapped from the parent Seurat object. Metadata about the sender cell in each cell-cell pair is prepended with "sender_" and metadata about the receiver cell prepended with "receiver_"
}
\description{
Map metadata from Seurat object to CCIM object
}
