% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapvalues}
\alias{mapvalues}
\title{Mapvalues (ripped from plyr)}
\usage{
mapvalues(x, from, to, warn_missing = TRUE)
}
\arguments{
\item{warn_missing}{}
}
\value{

}
\description{
Mapvalues (ripped from plyr)
}
