% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{PearsonDist}
\alias{PearsonDist}
\title{Plot Pearson distributions for ligand activities}
\usage{
PearsonDist(ligand_activities, pearson_cutoff = 0.075)
}
\arguments{
\item{ligand_activities}{the matrix output of \code{RankActiveLigands}}

\item{pearson_cutoff}{determines where to draw a vertical line illustrating the chosen Pearson threshold cutoff (default: 0.075)}
}
\value{
Two plots of all Pearson coefficients (left) and median Pearson coefficient per cell (right)
}
\description{
Plot Pearson distributions for ligand activities
}
