% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_diff.R
\name{PerturbedBins}
\alias{PerturbedBins}
\title{Identify most perturbed bins across >2 samples}
\usage{
PerturbedBins(seu, interaction_graphs = NULL)
}
\arguments{
\item{seu}{A binned Seurat object with bin identities in the "bin" column of meta.data}

\item{interaction_graphs}{List of summarized interaction graphs built from \code{BuildPriorInteraction} or \code{BuiltWeightedInteraction}}
}
\value{
For each sender bin-receiver bin combination, returns a data.frame of Kruskal-Wallis p-value and test statistic
}
\description{
Identify most perturbed bins across >2 samples
}
