% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_diff.R
\name{PerturbedBinSummary}
\alias{PerturbedBinSummary}
\title{Title}
\usage{
PerturbedBinSummary(seu, bin_pvals, cell.type.calls = "celltype.l2")
}
\arguments{
\item{cell.type.calls}{}
}
\value{

}
\description{
Title
}
