% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{PlotAlluviumLT}
\alias{PlotAlluviumLT}
\title{Plot Alluvial of ligand-target links}
\usage{
PlotAlluviumLT(connectome, optimize.flows = T, colors = NULL)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{
Plot Alluvial of ligand-target links
}
