% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_edges.R
\name{PlotLigandTargetAlluvium}
\alias{PlotLigandTargetAlluvium}
\title{Plot connections between ligands and predicted target genes
This function plots an alluvium displaying ligand-target linkages. If you consider a subset of \code{receiver_cells}, and a set of \code{ligands_of_interest}, perhaps those expressed by a particular sender cell type of interest, this function will find the top \code{n_targets} target genes predicted to be upregulated by those ligands. The only target genes shown are those that also appear in the individual cell's gene signature.}
\usage{
PlotLigandTargetAlluvium(
  seu,
  signature_matrix = NULL,
  active_ligands = active_ligands,
  receiver_cells = NULL,
  ligands_of_interest = NULL,
  n_targets = 250
)
}
\arguments{
\item{seu}{A seurat object}

\item{signature_matrix}{The output of \code{GenerateCellSignature}}

\item{active_ligands}{The output of \code{RankActiveLigands}}

\item{receiver_cells}{Character vector of receiver cells to query}

\item{ligands_of_interest}{Ligands to plot}

\item{n_targets}{Top ligand-target gene linkages for each ligand to plot}
}
\value{
A ggplot object
}
\description{
Plot connections between ligands and predicted target genes
This function plots an alluvium displaying ligand-target linkages. If you consider a subset of \code{receiver_cells}, and a set of \code{ligands_of_interest}, perhaps those expressed by a particular sender cell type of interest, this function will find the top \code{n_targets} target genes predicted to be upregulated by those ligands. The only target genes shown are those that also appear in the individual cell's gene signature.
}
