% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactome.R, R/viz.R
\name{PrioritizeInteractome}
\alias{PrioritizeInteractome}
\title{Title}
\usage{
PrioritizeInteractome(
  seu,
  gene_rankings,
  interactome,
  min.pct = 0.05,
  assay.use = "RNA",
  slot.use = "counts",
  send_cells = NULL,
  rec_cells = NULL
)

PrioritizeInteractome(
  seu,
  gene_rankings,
  interactome,
  min.pct = 0.05,
  assay.use = "RNA",
  slot.use = "counts",
  send_cells = NULL,
  rec_cells = NULL
)
}
\arguments{
\item{rec_cells}{}
}
\value{



}
\description{
Title

Title
}
