% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_edges.R
\name{PrioritizeLigands}
\alias{PrioritizeLigands}
\title{Title}
\usage{
PrioritizeLigands(
  seu,
  gene_rankings = NULL,
  min.pct = 0.025,
  assay = "RNA",
  slot = "counts"
)
}
\arguments{
\item{slot}{}
}
\value{

}
\description{
Title
}
