% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{random_connectivity_test}
\alias{random_connectivity_test}
\title{Test bin connectivity}
\usage{
random_connectivity_test(
  seu_oi = seu_oi,
  SNN = SNN,
  bin = bin,
  split.by = split.by,
  sigtest_cell_types = sigtest_cell_types
)
}
\arguments{
\item{seu_oi}{A Seurat object with bin identities in the "bins" column of the meta.data slot}

\item{SNN}{The SNN to use for connectivity testing}

\item{bin}{Character string of the bin ID to test}

\item{split.by}{Name of meta.data column containing the factor by which the dataset will be split}

\item{sigtest_cell_types}{Name of meta.data column containing the cell type labels to use for connectivity testing}
}
\value{
p-value of bin connectivity significance
}
\description{
Test bin connectivity
}
\examples{
random_connectivity_test()
}
