% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_edges.R
\name{RankActiveLigands}
\alias{RankActiveLigands}
\title{Predict ligand activity using cell-resolved gene signatures}
\usage{
RankActiveLigands(
  seu,
  variant_genes,
  dq = 0.5,
  species = "human",
  database = "OmniPath",
  ligands = NULL,
  recepts = NULL,
  ...
)
}
\arguments{
\item{seu}{A seurat object}

\item{variant_genes}{character vector of genes to use for gene-signature calculation. This list of genes should comprise genes that are variable across the biological axis of interest.
Eg. for single datasets, this may simply be the dataset's HVGs. For a longitudinal dataset, these may be genes that change over time. Use \code{IDVariantGenes} to identify genes that vary across time, between groups, etc.}

\item{dq}{numeric. Distance quantile. Genes that are further away than this quantile threshold will not be considered part of a cell's gene signature and not used for ligand activity prediction}

\item{species}{character. Name of species from which to load ligand-receptor databases. One of: "human", "mouse", "rat". Default: "human"}

\item{database}{Name of ligand-receptor database to use. Default: "OmniPath"
When species is "human", one of: OmniPath, CellChatDB, CellPhoneDB, Ramilowski2015, Baccin2019, LRdb, Kirouac2010, ICELLNET, iTALK, EMBRACE, HPMR, Guide2Pharma, connectomeDB2020, talklr, CellTalkDB
When species is "mouse" or "rat", only "OmniPath" is supported.
To pass a custom ligand-receptor database to this function, set database = "custom"}

\item{ligands}{Character vector of custom ligands to use for interaction graph generation. Ignored unless database = "custom"
When ligands is supplied, recepts must also be supplied and equidimensional.}

\item{recepts}{Character vector of custom receptors to use for interaction graph generation. Ignored unless database = "custom"
When recepts is supplied, ligands must also be supplied and equidimensional.}

\item{...}{Additional arguments passed to IDPotentialLigands}
}
\value{
Returns a matrix where columns are cells, rows are potential ligands, and values are pearson coefficients corresponding to each ligand's predicted activity in that cell.
}
\description{
Predict ligand activity using cell-resolved gene signatures
}
\references{
Browaeys, et al. Nat Methods (2019); Cortal, et al. Nat Biotech (2021)
}
