% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_edges.R
\name{RankLigandTargets}
\alias{RankLigandTargets}
\title{Predict ligand activity using cell-resolved gene signatures}
\usage{
RankLigandTargets(
  seu,
  signature_matrix,
  potential_ligands = NULL,
  species = "human",
  database = "OmniPath",
  receiver_cells = NULL,
  ligands = NULL,
  recepts = NULL,
  ntargets = 250,
  ...
)
}
\arguments{
\item{seu}{A seurat object}

\item{signature_matrix}{The output of \code{GenerateCellSignature}}

\item{potential_ligands}{character vector of ligands to include in active ligand ranking.}

\item{species}{character. Name of species from which to load ligand-receptor databases. One of: "human", "mouse", "rat". Default: "human"}

\item{database}{Name of ligand-receptor database to use. Default: "OmniPath"
When species is "human", one of: OmniPath, CellChatDB, CellPhoneDB, Ramilowski2015, Baccin2019, LRdb, Kirouac2010, ICELLNET, iTALK, EMBRACE, HPMR, Guide2Pharma, connectomeDB2020, talklr, CellTalkDB
When species is "mouse" or "rat", only "OmniPath" is supported.
To pass a custom ligand-receptor database to this function, set database = "custom"}

\item{ligands}{Character vector of custom ligands to use for interaction graph generation. Ignored unless database = "custom"
When ligands is supplied, recepts must also be supplied and equidimensional.}

\item{recepts}{Character vector of custom receptors to use for interaction graph generation. Ignored unless database = "custom"
When recepts is supplied, ligands must also be supplied and equidimensional.}

\item{ntargets}{The number of top targets from the ligand-target matrix to consider as active targets of that ligand}

\item{...}{Additional arguments passed to IDPotentialLigands}
}
\value{
Returns a matrix where columns are cells, rows are potential ligands, and values are pearson coefficients corresponding to each ligand's predicted activity in that cell.
}
\description{
Predict ligand activity using cell-resolved gene signatures
}
\references{
Browaeys, et al. Nat Methods (2019)
}
