% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_testing.R
\name{scInteraction}
\alias{scInteraction}
\title{Title}
\usage{
scInteraction(
  object,
  assay = "SCT",
  slot = "data",
  database = "OmniPath",
  ligands = NULL,
  recepts = NULL,
  specific = F,
  ranked_genes = NULL,
  cell.type.calls = "celltype.l2",
  dims_umap = 1:50,
  dims_neighbors = 1:50,
  cluster_resolution = 0.1
)
}
\arguments{
\item{cluster_resolution}{}
}
\value{

}
\description{
Title
}
