% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_diff.R
\name{ScoreBins}
\alias{ScoreBins}
\title{Assign quality scores to each bin}
\usage{
ScoreBins(seu, celltype.calls = NULL, split.by = NULL)
}
\arguments{
\item{seu}{A binned Seurat object with bin identities in the "bin" column of meta.data}

\item{celltype.calls}{Name of meta.data column containing the cell type labels to use for connectivity testing. Must be a finer (more granular) cell type label than used for coarse_cell_types in \code{BinDatasets}.}

\item{split.by}{Meta.data column name indicating how data was split for interaction graph generation}
}
\value{
A named numeric vector of bin quality scores
}
\description{
Assign quality scores to each bin
}
