% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_programs.R
\name{ScoreInteractionPrograms}
\alias{ScoreInteractionPrograms}
\title{Score expression of single-cells by expression of discovered interaction programs}
\usage{
ScoreInteractionPrograms(object, mods, return.assay = T)
}
\arguments{
\item{object}{A Seurat object to be scored}

\item{mods}{Interaction program data. Either the data.frame output of \code{InteractionProgramSignificance}, or a list of interaction program genes, as in the output of \code{InteractionPrograms} or \code{FindAllInteractionPrograms}}

\item{return.assay}{}
}
\value{
A Seurat object. When \code{return.assay = R}, returns interaction program ligand scores as an assay \code{IP_ligands} and receptor scores as an assay \code{IP_receptors}.
When \code{return.assay = F}, Returns a Seurat object with interaction program scores as meta.data columns. Ligand and receptor expression of interaction programs are scored separately.
The scores for sender cells (ligand expression) are stored in columns named "ligands_\link{interaction program name}"
The scores for receiver cells (receptor expression) are stored in columns named "receptors_\link{interaction program name}"
}
\description{
Score expression of single-cells by expression of discovered interaction programs
}
