% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elbow}
\alias{elbow}
\title{Detect Inflection Point of a Concave Curve (Elbox Method). Ripped from:
\url{https://github.com/ahasverus/elbow}}
\usage{
elbow(data)
}
\arguments{
\item{data}{a two-columns data frame (x and y respectively).}
}
\value{
This function returns a 2-elements list with:
\itemize{
\item the value on x-Axis corresponding to the inflection point
\item a data frame with the original data and two additional columns used in
the graphic.
}
}
\description{
This function implements the Elbow (or knee of a curve) method to detect
the inflection point of a concave curve. More information on this method:
\url{https://en.wikipedia.org/wiki/Elbow_method_(clustering)}.
}
\details{
This function detects unique inflection point in a simple concave
curve. The curve can be concave down/up with a positive/negative slope.
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.com}
}
