% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuits.R
\name{FindCircuits}
\alias{FindCircuits}
\title{Identify circuits between two timepoints}
\usage{
FindCircuits(
  seu,
  nnr,
  ranked_genes,
  tps,
  split.by = "orig.ident",
  assay.use = "SCT",
  slot.use = "data",
  pearson.cutoff = 0.075
)
}
\arguments{
\item{split.by}{}
}
\value{

}
\description{
Identify circuits between two timepoints
}
