% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_programs.R
\name{InteractionPrograms}
\alias{InteractionPrograms}
\title{Discover co-expressed ligand-receptor interaction programs}
\usage{
InteractionPrograms(
  object,
  assay = "SCT",
  slot = "data",
  species = "human",
  database = "OmniPath",
  ligands = NULL,
  recepts = NULL,
  iterate.threshold = 300,
  n.iterate = NULL,
  specific = F,
  ranked_genes = NULL,
  return.mat = T,
  softPower = NULL,
  r2_cutoff = 0.6,
  min.size = 5,
  plot.mods = F,
  tree.cut.quantile = 0.4,
  threads = NULL,
  cell_types = NULL,
  min.cell = 3
)
}
\arguments{
\item{object}{A seurat object}

\item{assay}{Assay in Seurat object from which to pull expression values}

\item{slot}{Slot within assay from which to pull expression values}

\item{species}{character. Name of species from which to load ligand-receptor databases. One of: "human", "mouse", "rat". Default: "human"}

\item{database}{Name of ligand-receptor database to use. Default: "OmniPath"
When species is "human", one of: OmniPath, CellChatDB, CellPhoneDB, Ramilowski2015, Baccin2019, LRdb, Kirouac2010, ICELLNET, iTALK, EMBRACE, HPMR, Guide2Pharma, connectomeDB2020, talklr, CellTalkDB
When species is "mouse" or "rat", only "OmniPath" is supported.
To pass a custom ligand-receptor database to this function, set database = "custom"}

\item{ligands}{Character vector of custom ligands to use for interaction graph generation. Ignored unless database = "custom"
When ligands is supplied, recepts must also be supplied and equidimensional.}

\item{recepts}{Character vector of custom receptors to use for interaction graph generation. Ignored unless database = "custom"
When recepts is supplied, ligands must also be supplied and equidimensional.}

\item{iterate.threshold}{When a dataset has more cells than \code{iterate.threshold}, the TOM will be approximated iteratively. For each iteration, a randomly subsampled dataset of size \code{iterate.threshold} will be used to construct the CCIM and generate the TOM.}

\item{n.iterate}{For datasets larger than \code{iterate.threshold}, determines how many iterations to perform to approximate the TOM.}

\item{specific}{logical. When TRUE, consider only the genes in each cell's predefined gene signature (see crGeneSig) as expressed. Default FALSE}

\item{ranked_genes}{Cell-resolved gene signatures, used only when specific = T}

\item{return.mat}{logical. Returns ligand-receptor covariance matrix and TOM along with modules and intramodular connectivity. Required for significance testing.}

\item{softPower}{softPower threshold for adjacency matrix. If unspecified, will be chosen automatically based on the minimum softPower that results in a scale-free topology fitting index (R^2) of greater than r2_cutoff (by default 0.6)}

\item{r2_cutoff}{The softPower will be chosen as the minimum value that satisfies a scale-free topology fitting index (R^2) of r2_cutoff (by default: 0.6)}

\item{min.size}{Minimum size of each interaction program}

\item{plot.mods}{Plot modules and associated dendrograms during analysis}

\item{tree.cut.quantile}{The dendrogram tree height quantile at which the dendrogram should be cut. Higher values lead to fewer, smaller modules.}

\item{threads}{To enable WGCNA multi-threading, specify the number of threads to allow. This parameter may be required when running on multiple cores}

\item{cell_types}{\code{meta.data} column name corresponding to cell types or clusters. If iteratively approximating the TOM, when specified the sequences of subsampled CCIM will be sampled proportionally to annotations present in this grouping.}

\item{min.cell}{When \code{cell_types} is specified, the minimum number of cells in a cell type that will be included when generating each iterative CCIM (default: 3)}
}
\value{
When return.mat = T, returns a list of length 4 containing ligand-receptor covariance matrix, TOM, module lists, and intramodular connectivity. Otherwise, returns a list of length 2 containing only module lists and intramodular connectivity.
}
\description{
Discover co-expressed ligand-receptor interaction programs
}
