% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapvalues}
\alias{mapvalues}
\title{Mapvalues (ripped from plyr)}
\usage{
mapvalues(x, from, to, warn_missing = TRUE)
}
\arguments{
\item{x}{the factor or vector to modify}

\item{from}{a vector of the items to replace}

\item{to}{a vector of replacement values}

\item{warn_missing}{print a message if any of the old values are not present in x}
}
\value{

}
\description{
Mapvalues (ripped from plyr)
}
